[gpio] Move mock code into a separate zjs_gpio_mock.c file
Also, add ZJS_FLAGS variable to Makefile to pass command line
flags to ZJS build. This lets you manually enable -DZJS_GPIO_MOCK
to run mock gpio build on A101 hardware for example.
